feat: refactored itest code; added objecttypes token to itest
