+++ docker-compose.yaml
@@ -384,7 +384,7 @@ services:
       - OBJECTS_API_CLIENT_MP_REST_URL=http://objecten-api:8000
       - OBJECTS_API_TOKEN=1 # dummy for now
       - OBJECTTYPES_API_CLIENT_MP_REST_URL=http://objecttypen-api:8000
-      - OBJECTTYPES_API_TOKEN=1 # dummy for now
+      - OBJECTTYPES_API_TOKEN=dummyzacobjecttypestoken
       - OPA_API_CLIENT_MP_REST_URL=http://opa:8181
       - OPEN_FORMS_URL=http://localhost:9999 # dummy for now
       - OPEN_NOTIFICATIONS_API_SECRET_KEY=opennotificaties
